Watchdog: Refactored module structure
