+++ arch/cortex-m/samv71-hal/src/drivers/watchdog/watchdog_config.rs
